Add default code of conduct
This code of conduct is shown for all repositories wich don't have
there own CODE_OF_CONDUCT.md file. This uses a GitHub feature documented at
https://help.github.com/en/articles/creating-a-default-community-health-file-for-your-organization.

For the main repos we should still have the file in the repo so that
it's also available out of the context of GitHub, but having a default
file makes sure that there always is one available, also for new
repositorires or smaller ones.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>